user-requirements.txt
mkdocs
mkdocs-material
mkdocs-macros-plugin
mkdocs-section-index
pyyaml
